+++ src/main.h
@@ -1,5 +1,5 @@
-﻿//	Error Lookup
-//	Copyright © 2015 Henry++
+﻿// Error Lookup
+// Copyright © 2011, 2012, 2015 Henry++
 
 #ifndef __MAIN_H__
 #define __MAIN_H__
@@ -26,6 +26,24 @@
 #define STATUS_SEVERITY_WARNING			0x2
 #define STATUS_SEVERITY_ERROR			0x3
 
+#define FACILITY_AUDIO_KERNEL			68
+#define FACILITY_BTH_ATT                 0x42
+#define FACILITY_CODCLASS_ERROR_CODE     0x6
+#define FACILITY_FILTER_MANAGER          0x1C
+#define FACILITY_GRAPHICS_KERNEL         0x1E
+#define FACILITY_INTERIX                 0x99
+#define FACILITY_MAXIMUM_VALUE           0xE8
+#define FACILITY_MONITOR                 0x1D
+#define FACILITY_WIN32K_NTGDI            0x3F
+#define FACILITY_WIN32K_NTUSER           0x3E
+#define FACILITY_RDBSS                   0x41
+#define FACILITY_RESUME_KEY_FILTER       0x40
+#define FACILITY_SECUREBOOT              0x43
+#define FACILITY_SHARED_VHDX             0x5C
+#define FACILITY_SPACES                  0xE7
+#define FACILITY_VIDEO                   0x1B
+#define FACILITY_VOLSNAP                 0x50
+
 #define FACILITY_D3D					0x876
 #define FACILITY_D3DX					0x877
 #define FACILITY_DSOUND_DMUSIC			0x878
@@ -38,6 +56,8 @@
 #define FACILITY_D3D11_OR_AE			0x87c
 #define FACILITY_APO					0x87d
 #define FACILITY_WSAPI					0x889
+
+#define FACILITY_EAP_MESSAGE           0x0842
 
 #define SWITCH_ROUTINE(x)  case x: StringCchPrintf(buffer, MAX_PATH, L"%s (0x%02x)\0", L#x, x); break;